fixed from PVS-Studio:
V528 It is odd that pointer to 'wchar_t' type is compared with the L'\0' value. Probably meant: *cmdLinePastCompList == L'\0'. parsing.cpp 335
V618 It's dangerous to call the 'printf' function in such a manner, as the line being passed could contain format specification. The example of the safe code: printf("%s", str); paexec.cpp 334